Eco-Sort/requirements.txt
streamlit
numpy
pandas
pillow
tensorflow
urllib3